move other common data types into common
