+++ lib/streamlit/runtime/state/widgets.py
@@ -21,11 +21,6 @@ from typing_extensions import Final, TypeAlias
 from streamlit.errors import DuplicateWidgetID
 from streamlit.proto.WidgetStates_pb2 import WidgetState, WidgetStates
 from streamlit.runtime.state.common import (
-    WidgetProto,
-    compute_widget_id,
-    user_key_from_widget_id,
-)
-from streamlit.runtime.state.session_state import (
     RegisterWidgetResult,
     T,
     WidgetArgs,
@@ -33,7 +28,10 @@ from streamlit.runtime.state.session_state import (
     WidgetDeserializer,
     WidgetKwargs,
     WidgetMetadata,
+    WidgetProto,
     WidgetSerializer,
+    compute_widget_id,
+    user_key_from_widget_id,
 )
 from streamlit.type_util import ValueFieldName
